create changeData.js
